requirements/dev.txt
-r common.txt
grip==4.5.2
factory-boy==2.11.1
pytest-cov==2.6.0
pytest-django==3.4.3
